+++ packages/ax-code/src/provider/ax-engine/index.ts
@@ -1,4 +1,6 @@
 export * from "./constants"
+export * from "./bundled"
+export * from "./payload"
 export * from "./catalog"
 export * from "./connection"
 export * from "./delete"